Added New Employee Feature
In this fundamental commit, the classes and interfaces files have been
created according to the design developed. Connection to persistent data
is made with the help of Json.simple library. If time permits, this will
be converted to the more reliable SQL database.
This commit also covers the "display short employee details" feature,
which fetches the employee details from the stored .json file and parses
it to stdout (console).