Added Delete Employee Feature
Modified the design of the classes and added the 'delete employee'
feature. In the prev commit, HourlyEmp and MonthlyEmp were deriving from
the main Emp class. Now, those have been altered to composition of
HourlyPayDetails and MonthlyPayDetails inside the Emp class. Also,
persistent data with JSON will take a longer time to execute. Hence, to
meet the deadline, decided to use non-persistent data.